refactor: 보드 초기화 방식 인터페이스화
